Fixed ref issue and minor problems with HELLO's
The ref issue has been tested (not thoroughly, but all the components
appear to have the correct values) and seems to work now.
Fixed an issue where a message was not being serialised correctly.
Fixed and issue in the HELLO handler where it was not inserting all of the node
elements.
Removed the ISocketHandler.AddMessageHandler method.
General layout changes.