Added responseManager
Added a class to handle tracking messages which require a response from
the server (messages that require an ACK response).
Added new settings for this class
Added it to the DataManager and to PTPClient
Renamed Classes/Message to ChatMessage (im happier with this one).

It still needs to be piped into SocketHandler and ChannelVerbHandler but
its 10 to two, I'm tired