Plumbed in ResponseManager to SocketHandler
Had to change the Interface for ResponseManager as SocketHandler needed
more info about the message. ResponseMessage is now passed out of the
GetMessages and GetOutstandingMessages methods.
Continued work on the ChannelHandler, which now handles sending a connect
to unknownNodes, if we are part of the channel described in the message.