Started work on ChannelTab
ChannelTab now has abit more going on, Has a DataListView with a textbox
and button for sending messages to that channel.
There is a new interface, for the channel tab, to encapsulate the
interaction between the tab and other components.
PtpClient also grabs some of the channel event elements as well.
For testing, I have added a default channel which PtpClient creates.