Few changes
Protocol has changed to put message id in the root of the message.
Updated BaseVerbHandler to handle most (if not all) of the root message
elements.
Updated RoutingVerbHandler to better handle TTL values.
Updated HelloVerbHandler to better handle TTL values.
Edited the Node const, to require the node id there.
Testing the use of Generic types in MessageHandler.
Added a TryParseNodeId to ExtensionMethods.
Added ToString and Equals methods to NodeId.