Added EventHandlers to SocketHandler and thread.
These event handlers can be used to be notified when a socket
connects/closes/reset. The the moment, this is used to fire off a hello
message when the socket is connected.

Reluctantly, I have also added the json package to the Base package. this
is so that the SenderId and TargetId are ignored (by assigning json
attributes to them) when the class is being serialised.